v0.3.0:#19: Add a CLI option to enable to use a custom UI: -c, --custom-ui.
If -c /path/to/ui_script.py is provided:
dirarchy creates a temp file which is provided to ui_script.py.
After the script execution, dirarchy uses the temp file filled by ui_script.py to extract the variables values.

Variables are expected to be organized in a JSON dict.